refactor(spot): split up into multiple components
re #146
