+++ packages/cms-spa/src/components/spot/Spot-Banner.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { Container } from '../Container';
+export function Banner(message: string) {
+  return (
+    <div className='notification spot__calib-notification--on-top'>
+      {message}
+    </div>
+  );
+}
