+++ packages/cms-spa/src/components/spot/Spot-Hr.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { Container } from '../Container';
+export function SpotHr() {
+  return (
+    <Container>
+      <hr />
+    </Container>
+  );
+}
