+++ packages/cms-spa/src/components/spot/Spot-TableTitle.tsx
@@ -1,0 +1,42 @@
+import React from 'react';
+import { IconRain, IconCalc, IconComment, IconCSV } from '../fontawesome-icons';
+export type TableTitleIconType =
+  | 'IconCalc'
+  | 'IconRain'
+  | 'IconCSV'
+  | 'IconComment';
+export interface ISpotTableTitle {
+  title: string;
+  iconType: TableTitleIconType;
+}
+export function SpotTableTitle({
+  title,
+  iconType,
+}: ISpotTableTitle): React.ReactNode {
+  return (
+    <h3 className='is-title is-3'>
+      <span>
+        {(() => {
+          switch (iconType) {
+            case 'IconComment': {
+              return <IconComment></IconComment>;
+            }
+            case 'IconCSV': {
+              return <IconCSV></IconCSV>;
+            }
+            case 'IconRain': {
+              return <IconRain></IconRain>;
+            }
+            case 'IconCalc': {
+              return <IconCalc></IconCalc>;
+            }
+            default: {
+              throw new Error('No default icon defiend');
+            }
+          }
+        })()}
+      </span>{' '}
+      <span>{title}</span>
+    </h3>
+  );
+}
